Additional fixes were made due to Judge's failed unit testing

+++ Sandbox Solutions/EF Core/EntityRelations/P02_FootballBetting.Data.Models/Game.cs
@@ -27,10 +27,10 @@ namespace P02_FootballBetting.Data.Models
         public Team AwayTeam { get; set; } = null!;
 
         [Required]
-        public sbyte HomeTeamGoals { get; set; }
+        public int HomeTeamGoals { get; set; }
 
         [Required]
-        public sbyte AwayTeamGoals { get; set; }
+        public int AwayTeamGoals { get; set; }
 
         [Required]
         public double HomeTeamBetRate { get; set; }
