+++ Sandbox Solutions/EF Core/EntityRelations/P02_FootballBetting.Data.Models/PlayerStatistic.cs
@@ -20,12 +20,12 @@ namespace P02_FootballBetting.Data.Models
         public virtual Player Player { get; set; } = null!;
 
         [Required]
-        public sbyte ScoredGoals { get; set; }
+        public int ScoredGoals { get; set; }
 
         [Required]
-        public sbyte Assists { get; set; }
+        public int Assists { get; set; }
 
         [Required]
-        public ushort MinutesPlayed { get; set; }
+        public int MinutesPlayed { get; set; }
     }
 }
